fix: Extend transfer reason cleanup regexp with all allowed characters
The field accepts the same characters as recipient label. The mobile
application accepts a few more characters but stick to web application
list to be safe. The expressions are simplified to avoid unnecessary
escaping. Only the \- is preserved to keep special characters grouped as
in the applications, otherwise it would need to be moved to the start of
the group.