style: simplify attribute checking
flake8 is complaining that getattr of a constant is no safer than direct
use of the attribute, which is true here since the third argument
to make it safe is missing. This is not a big problem though as
the attribute presence is checked above but there is also
an AttributeNotFound exception handling afterwards which covers this.
Move the attribute access in the try/except and remove unnecessary code.