chore: add type annotations to woob.tools.decorators
It is currently not possible to use ParamSpec and inject new optional
keyword arguments. This is unfortunate but it should not be a problem as
the feature to inject `_tries` or `_delay` at function call site does
not appear to be used anywhere. Choosing to favor full type checking
over supporting a rarely used feature for now.

Refs: https://stackoverflow.com/questions/79679286/how-to-use-a-decorator-to-mark-a-keyword-only-mandatory-argument-as-optional-in