chore: add type annotations to woob.tools.regex_helper
Change fallback value for Group on line 185 and `get_quantifier` to an
empty string in order to avoid type inconsistencies and complicated
workarounds. This change should have no effect as empty string also falsy
in the tests where None was used and does not generate unwanted text
output in str concatenations.